Conversion to spring/hibernate project

+++ src/main/java/happy/rabbit/domain/JenkinsItem.java
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
 
 public class JenkinsItem {
 
-    private Integer id;
+    private Long id;
     private FailureReason failureReason;
     private String content;
     private boolean isBroken;
@@ -17,17 +17,17 @@ public class JenkinsItem {
 
     }
 
-    public JenkinsItem(int id, FailureReason failureReason, String content) {
+    public JenkinsItem(Long id, FailureReason failureReason, String content) {
         this.id = id;
         this.failureReason = failureReason;
         this.content = content;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
